move JEI plugin to deps subfolder to prevent unintended loading
